Moved JUnits to Main to ease discussion

+++ 2.JPA/non-project/src/main/java/edu/depaul/se/jpa/basic/createdtables/BookService.java
@@ -38,7 +38,7 @@ public class BookService {
         return em.createNamedQuery("findAllBooks").getResultList();
     }
 
-    public void deleteBook(int bookId) {
+    public void deleteBook(long bookId) {
         EntityTransaction tx = em.getTransaction();
 
         tx.begin();
